Change init/shutdown tests to be consistent
This package's init function has a hot path when it has already been
initialized, so the first iteration of the init test was cold and the
rest were hot. I changed that test to specifically test
re-initialization by performing the cold init outside of the timing
loop. That test also wasn't previously shutting down when the run was
completed.

I added a cold initialize call to the shutdown test, so that test should
give us the full startup/shutdown times now. So the two tests now give
us an idea of what the first startup is like, as well as subsequent
"hot" startups.

Signed-off-by: Scott K Logan <logans@cottsay.net>