Fix heap allocation statistics
For the log-hit test, there must be some heap allocation associated with
the first call, because we aren't getting a consistent number of
allocations on that first call. We can resolve that by explicitly
"priming" the logger, then timing only the consistently behaving calls.

For the re-initialize test, of course the cold startup is performing
more allocations than the hot startup, so we need to exclude those
allocations from the metric.

Signed-off-by: Scott K Logan <logans@cottsay.net>