Update CMakeLists.txt to use C++17 (#76)
The newest version of log4cxx on macOS is now built with C++17 on homebrew (the expected way to install the package, based on the ROS2 installation guide). C++17 adds support for shared_mutex, so log4cxx defaults to using shared_mutex in std instead of boost. C++14, however, doesn't have shared_mutex, so building rcl_logging_log4cxx on macOS with the latest packages now fails because it uses C++14. The easiest way to fix this is to switch this package to build on C++17, which is what I've done here.

Signed-off-by: nkalupahana <avishkank@gmail.com>